DRY analyzer specs
These pieces were repeated pretty much verbatim between the different
analyzer specs. That seems especially silly since this is, after all, an
engine for pointing out code duplication.
