+++ spec/cc/engine/analyzers/php/main_spec.rb
@@ -5,16 +5,8 @@ require 'cc/engine/analyzers/file_list'
 require 'flay'
 require 'tmpdir'
 
-RSpec.describe CC::Engine::Analyzers::Php::Main do
-  around do |example|
-    Dir.mktmpdir do |directory|
-      @code = directory
-
-      Dir.chdir(directory) do
-        example.run
-      end
-    end
-  end
+RSpec.describe CC::Engine::Analyzers::Php::Main, in_tmpdir: true do
+  include AnalyzerSpecHelpers
 
   describe "#run" do
     it "prints an issue" do
@@ -56,21 +48,6 @@ RSpec.describe CC::Engine::Analyzers::Php::Main do
       expect(json["content"]["body"]).to match /This issue has a mass of `44`/
       expect(json["fingerprint"]).to eq("667da0e2bab866aa2fe9d014a65d57d9")
     end
-  end
-
-  def create_source_file(path, content)
-    File.write(File.join(@code, path), content)
-  end
-
-  def run_engine(config = nil)
-    io = StringIO.new
-
-    engine = ::CC::Engine::Analyzers::Php::Main.new(engine_config: config)
-    reporter = ::CC::Engine::Analyzers::Reporter.new(double(concurrency: 2), engine, io)
-
-    reporter.run
-
-    io.string
   end
 
   def printed_issue
